ColorUtils
refactor
